refactor(payments): Unify LangGraph and Strands config into single class
  Merge AgentCorePaymentsConfig (LangGraph) and AgentCorePaymentsPluginConfig
  (Strands) into a single dataclass in integrations/config.py. Both names
  remain available as aliases for backward compatibility.
