+++ src/bedrock_agentcore/payments/integrations/langgraph/__init__.py
@@ -1,6 +1,6 @@
 """LangGraph integration for AgentCore Payments."""
 
-from .config import AgentCorePaymentsConfig
+from ..config import AgentCorePaymentsConfig
 from .errors import ErrorResolution, PaymentErrorContext
 from .middleware import AgentCorePaymentsMiddleware
