+++ tests/bedrock_agentcore/payments/integrations/langgraph/test_stage5.py
@@ -8,7 +8,7 @@ import pytest
 from langchain.messages import ToolMessage
 from langgraph.types import Command
 
-from bedrock_agentcore.payments.integrations.langgraph.config import AgentCorePaymentsConfig
+from bedrock_agentcore.payments.integrations.langgraph import AgentCorePaymentsConfig
 from bedrock_agentcore.payments.integrations.langgraph.middleware import AgentCorePaymentsMiddleware
